requirements.txt
diffusers==0.29.2
streamlit==1.36.0
torch==2.3.1
diffusers==0.29.2
pillow==10.4.0
bcrypt==4.2.0
requests~=2.32.3
uvicorn~=0.30.3
fastapi~=0.111.1
sqlalchemy~=2.0.31
pydantic~=2.8.2